chore(otel): refactor crate
