+++ codex-rs/otel/src/events/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod session;
+pub(crate) mod shared;
