+++ codex-rs/otel/src/events/shared/trace.rs
@@ -1,0 +1,19 @@
+macro_rules! trace_event {
+    ($self:expr, $($fields:tt)*) => {{
+        tracing::event!(
+            target: $crate::targets::OTEL_TRACE_SAFE_TARGET,
+            tracing::Level::INFO,
+            $($fields)*
+            event.timestamp = %$crate::events::shared::timestamp(),
+            conversation.id = %$self.metadata.conversation_id,
+            app.version = %$self.metadata.app_version,
+            auth_mode = $self.metadata.auth_mode,
+            originator = %$self.metadata.originator,
+            terminal.type = %$self.metadata.terminal_type,
+            model = %$self.metadata.model,
+            slug = %$self.metadata.slug,
+        );
+    }};
+}
+
+pub(crate) use trace_event;
