--- codex-rs/otel/src/traces/mod.rs
@@ -1,1 +1,0 @@
-pub mod otel_manager;
